feat: 엔티티 ERD 반영 과정 수정
