+++ src/main/java/ject/petfit/domain/note/entity/Note.java
@@ -12,6 +12,7 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import java.time.LocalDate;
 import ject.petfit.domain.member.entity.Member;
+import ject.petfit.domain.pet.entity.Pet;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -40,18 +41,18 @@ public class Note {
     private LocalDate createdAt;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
+    @JoinColumn(name = "pet_id", nullable = false)
+    private Pet pet;
 
     @Builder
-    public Note(String title, String content, LocalDate createdAt, Member member) {
+    public Note(String title, String content, LocalDate createdAt, Pet pet) {
         this.title = title;
         this.content = content;
         this.createdAt = createdAt;
-        this.member = member;
+        this.pet = pet;
     }
 
-    public void setMember(Member member) {
-        this.member = member;
+    public void setPet(Pet pet) {
+        this.pet = pet;
     }
 }
